Add: basic countdown timer in seconds
